# docs/swagger.yaml
definitions:
  AuthOutput:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/UserOutput'
    type: object
  ClientError:
    properties:
      error:
        allOf:
        - $ref: '#/definitions/ErrCode'
        example: SERVER_NOT_RESPONDING
    type: object
  DeleteReminderOutput:
    properties:
      id:
        type: integer
    type: object
  ErrCode:
    enum:
    - SERVER_INTERNAL_ERROR
    - SERVER_NOT_RESPONDING
    - SERVICE_TEMPORARILY_UNAVAILABLE
    - USER_INVALID_DATA
    - USER_ALREADY_EXISTS
    - USER_NOT_FOUND
    - Reminder_INVALID_DATA
    - AUTH_INVALID_TOKEN
    - AUTH_EXPIRED_TOKEN
    - AUTH_TOKEN_NOT_PROVIDED
    - AUTH_INVALID_EMAIL_OR_PASSWORD
    type: string
    x-enum-varnames:
    - ErrServerInternalError
    - ErrServerNotResponding
    - ErrServiceTemporarilyUnavailable
    - ErrUserInvalidData
    - ErrUserAlreadyExists
    - ErrUserNotFound
    - ErrReminderInvalidData
    - ErrAuthInvalidToken
    - ErrAuthExpiredToken
    - ErrAuthTokenNotProvided
    - ErrAuthInvalidEmailOrPassword
  GetMeOutput:
    properties:
      user:
        $ref: '#/definitions/UserOutput'
    type: object
  GoogleOAuthOutput:
    properties:
      access_token:
        type: string
      user:
        $ref: '#/definitions/github_com_Toppira-Official_Reminder_Server_internal_modules_auth_usecase.GoogleUserInfo'
    type: object
  HttpOutput-AuthOutput:
    properties:
      data:
        $ref: '#/definitions/AuthOutput'
    type: object
  HttpOutput-DeleteReminderOutput:
    properties:
      data:
        $ref: '#/definitions/DeleteReminderOutput'
    type: object
  HttpOutput-GetMeOutput:
    properties:
      data:
        $ref: '#/definitions/GetMeOutput'
    type: object
  HttpOutput-GoogleOAuthOutput:
    properties:
      data:
        $ref: '#/definitions/GoogleOAuthOutput'
    type: object
  HttpOutput-MyRemindersOutput:
    properties:
      data:
        $ref: '#/definitions/MyRemindersOutput'
    type: object
  HttpOutput-NewReminderOutput:
    properties:
      data:
        $ref: '#/definitions/NewReminderOutput'
    type: object
  HttpOutput-UpdateMeAcceptedOutput:
    properties:
      data:
        $ref: '#/definitions/UpdateMeAcceptedOutput'
    type: object
  HttpOutput-UpdateMeOutput:
    properties:
      data:
        $ref: '#/definitions/UpdateMeOutput'
    type: object
  LoginWithEmailPasswordInput:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: StrongPassword1234
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  MyRemindersOutput:
    properties:
      reminders:
        items:
          $ref: '#/definitions/github_com_Toppira-Official_Reminder_Server_internal_shared_entities.Reminder'
        type: array
    type: object
  NewReminderInput:
    properties:
      description:
        type: string
      priority:
        $ref: '#/definitions/ReminderPriority'
      reminder_times:
        items:
          type: string
        type: array
      scheduled_at:
        type: string
      title:
        maxLength: 200
        minLength: 3
        type: string
    required:
    - scheduled_at
    - title
    type: object
  NewReminderOutput:
    properties:
      reminder:
        $ref: '#/definitions/github_com_Toppira-Official_Reminder_Server_internal_shared_entities.Reminder'
    type: object
  ReminderPriority:
    enum:
    - low
    - medium
    - high
    - critical
    type: string
    x-enum-varnames:
    - Low
    - Medium
    - High
    - Critical
  SignUpWithEmailPasswordInput:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: StrongPassword1234
        maxLength: 72
        minLength: 8
        type: string
    required:
    - email
    - password
    type: object
  UpdateMeAcceptedOutput:
    properties:
      message:
        type: string
    type: object
  UpdateMeInput:
    properties:
      name:
        example: John Doe
        type: string
      password:
        example: securepassword
        maxLength: 72
        minLength: 8
        type: string
      phone:
        example: "09123456789"
        type: string
    type: object
  UpdateMeOutput:
    properties:
      user:
        $ref: '#/definitions/UserOutput'
    type: object
  UserOutput:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      updated_at:
        type: string
    type: object
  github_com_Toppira-Official_Reminder_Server_internal_modules_auth_usecase.GoogleUserInfo:
    properties:
      email:
        type: string
      family_name:
        type: string
      name:
        type: string
      picture:
        type: string
    type: object
  github_com_Toppira-Official_Reminder_Server_internal_shared_constants.ReminderStatus:
    enum:
    - pending
    - completed
    - missed
    type: string
    x-enum-varnames:
    - Pending
    - Completed
    - Missed
  github_com_Toppira-Official_Reminder_Server_internal_shared_entities.Reminder:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      priority:
        $ref: '#/definitions/ReminderPriority'
      reminder_times:
        items:
          type: string
        type: array
      scheduled_at:
        type: string
      status:
        $ref: '#/definitions/github_com_Toppira-Official_Reminder_Server_internal_shared_constants.ReminderStatus'
      title:
        type: string
      updated_at:
        type: string
      user:
        $ref: '#/definitions/github_com_Toppira-Official_Reminder_Server_internal_shared_entities.User'
      user_id:
        type: integer
    type: object
  github_com_Toppira-Official_Reminder_Server_internal_shared_entities.User:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      phone:
        type: string
      profile_picture:
        type: string
      reminders:
        items:
          $ref: '#/definitions/github_com_Toppira-Official_Reminder_Server_internal_shared_entities.Reminder'
        type: array
      updated_at:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    email: AliMoradi0Business@gmail.com
    name: Ali Moradi
  description: Toppira APIs Documents.
  title: Toppira APIs
  version: "1.0"
paths:
  /auth/google-oauth/callback:
    get:
      parameters:
      - description: Code
        in: query
        name: code
        required: true
        type: string
      - description: State
        in: query
        name: state
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput-GoogleOAuthOutput'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
      summary: Handle Google OAuth callback
      tags:
      - Authentication
  /auth/google-oauth/redirect-url:
    get:
      responses:
        "307":
          description: Temporary Redirect
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
      summary: Redirect to Google OAuth URL
      tags:
      - Authentication
  /auth/login-with-user-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Login Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/LoginWithEmailPasswordInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput-AuthOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      summary: login with email and password
      tags:
      - Authentication
  /auth/sign-up-with-user-password:
    post:
      consumes:
      - application/json
      parameters:
      - description: Sign Up Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/SignUpWithEmailPasswordInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/HttpOutput-AuthOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      summary: sign up with email and password
      tags:
      - Authentication
  /reminder:
    get:
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput-MyRemindersOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      security:
      - BearerAuth: []
      summary: returns user's all reminders
      tags:
      - Reminder
    post:
      consumes:
      - application/json
      parameters:
      - description: New Reminder Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewReminderInput'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/HttpOutput-NewReminderOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      security:
      - BearerAuth: []
      summary: creates new reminder
      tags:
      - Reminder
  /reminder/{id}:
    delete:
      parameters:
      - description: Reminder ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput-DeleteReminderOutput'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      security:
      - BearerAuth: []
      summary: deletes the given reminder
      tags:
      - Reminder
  /user/me:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput-GetMeOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      security:
      - BearerAuth: []
      summary: get my(user) data
      tags:
      - User
    patch:
      consumes:
      - application/json
      parameters:
      - description: Update Me Input
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/UpdateMeInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/HttpOutput-UpdateMeOutput'
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/HttpOutput-UpdateMeAcceptedOutput'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/ClientError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/ClientError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/ClientError'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/ClientError'
      security:
      - BearerAuth: []
      summary: update my(user) data
      tags:
      - User
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"