# internal/modules/user/handler/get_me.go
package handler

import (
	"net/http"

	"github.com/Toppira-Official/Reminder_Server/internal/modules/user/handler/dto"
	output "github.com/Toppira-Official/Reminder_Server/internal/shared/dto"
	"github.com/Toppira-Official/Reminder_Server/internal/shared/entities"
	apperrors "github.com/Toppira-Official/Reminder_Server/internal/shared/errors"
	"github.com/gin-gonic/gin"
)

type GetMeHandler struct{}

func NewGetMeHandler() *GetMeHandler {
	return &GetMeHandler{}
}

// GetMyInfo godoc
//
//	@Summary	get my(user) data
//	@Tags		User
//	@Accept		json
//	@Produce	json
//	@Success	200	{object}	output.HttpOutput[dto.GetMeOutput]
//	@Failure	400	{object}	apperrors.ClientError
//	@Failure	401	{object}	apperrors.ClientError
//	@Failure	500	{object}	apperrors.ClientError
//	@Failure	503	{object}	apperrors.ClientError
//	@Security	BearerAuth
//	@Router		/user/me [get]
func (hl *GetMeHandler) GetMyInfo(c *gin.Context) {
	userVal, exists := c.Get("user")
	if !exists {
		c.Error(apperrors.E(apperrors.ErrUserNotFound))
		return
	}

	user, ok := userVal.(*entities.User)
	if !ok {
		c.Error(apperrors.E(apperrors.ErrUserNotFound))
		return
	}

	c.JSON(http.StatusOK, output.HttpOutput[dto.GetMeOutput]{
		Data: dto.GetMeOutput{
			User: output.ToUserOutput(user),
		},
	})
}